Improve the access_methods_from_raw_vector function

+++ ios/MullvadRustRuntime/MullvadConnectionModeProvider.swift
@@ -92,7 +92,7 @@ public func initAccessMethodSettingsWrapper(methods: [PersistentAccessMethod])
                 bridgesMethodRaw,
                 encryptedDNSMethodRaw,
                 $0.baseAddress!,
-                customMethods.count
+                UInt(customMethods.count)
             )
         }
     )
